Use raw counters instead of combined events

+++ src/hardware-counters/papi/PAPICPUHardwareCounters.hpp
@@ -8,6 +8,7 @@
 #define PAPI_CPU_HARDWARE_COUNTERS_HPP
 
 #include "hardware-counters/CPUHardwareCountersInterface.hpp"
+#include "hardware-counters/SupportedHardwareCounters.hpp"
 
 
 class PAPICPUHardwareCounters : public CPUHardwareCountersInterface {
@@ -18,7 +19,10 @@ public:
 	{
 	}
 
-	inline double getDelta(HWCounters::counters_t) override
+	//! \brief Get the delta value of a HW counter
+	//!
+	//! \param[in] counterType The type of counter to get the delta from
+	inline uint64_t getDelta(HWCounters::counters_t) override
 	{
 		return 0.0;
 	}
